Add FileResponderDataSourceImpl and make it the default responder
