+++ src/main/java/com/eighthlight/fabrial/http/FileResponderDataSourceImpl.java
@@ -1,0 +1,10 @@
+package com.eighthlight.fabrial.http;
+
+import java.nio.file.Path;
+
+public class FileResponderDataSourceImpl implements FileHttpResponder.DataSource {
+  @Override
+  public boolean fileExistsAtPath(Path path) {
+    return path.toFile().exists();
+  }
+}
